refactor:  cloudinary signed-upload controller and router
