+++ finpro-api/src/controllers/product.controller/product.controller.ts
@@ -1,6 +1,6 @@
 import { prisma } from "../../prisma";
-import { ICloudinaryResult, IProduct, IProductImage } from "../../types/product.type";
-import e, { NextFunction, Request, Response } from "express";
+import { ICloudinaryResult, IProductImage } from "../../types/product.type";
+import { NextFunction, Request, Response } from "express";
 import { v2 as cloudinary } from "cloudinary";
 import { Prisma } from "@prisma/client";
 import { EditProductImageService } from "../../services/product.service/product.image.service";
@@ -239,24 +239,6 @@ export class ProductController {
         success: true,
         message: "Product updated successfully",
         resultProduct,
-      });
-    } catch (error) {
-      next(error);
-    }
-  }
-
-  async handleSignedupload(req: Request, res: Response, next: NextFunction) {
-    try {
-      cloudinary.config({
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-        api_key: process.env.CLOUDINARY_API_KEY,
-        api_secret: process.env.CLOUDINARY_API_SECRET,
-        secure: true,
-      });
-      const { paramsToSign } = req.body;
-      const signature = cloudinary.utils.api_sign_request(paramsToSign, process.env.CLOUDINARY_API_SECRET as string);
-      res.status(200).json({
-        signature,
       });
     } catch (error) {
       next(error);
