+++ finpro-api/src/routers/cloudinary.router.ts
@@ -1,0 +1,7 @@
+import { handleSignedupload } from "../controllers/cloudinary.controller";
+
+const cloudinaryRouter = require("express").Router();
+
+cloudinaryRouter.post('/signed-upload',handleSignedupload)
+
+export default cloudinaryRouter;
