+++ finpro-web/src/features/admin/products/components/ProductImageUploadWidget.tsx
@@ -15,7 +15,7 @@ export default function ProductImageUploadWidget(props: IProductImageUploadButto
     <>
       <CldUploadWidget
         uploadPreset='products-image'
-        signatureEndpoint={`${process.env.NEXT_PUBLIC_API_BASE_URL}/product/signed-upload`}
+        signatureEndpoint={`${process.env.NEXT_PUBLIC_API_BASE_URL}/cloudinary/signed-upload`}
         onSuccess={props.onSuccess}
         options={{
           sources: ["local", "url", "camera", "image_search", "google_drive"],
